crits_next/backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
authlib
sqlmodel
